face detected, comic chosen, but no drawing
