+++ app/src/main/java/com/example/aounl/comicfaceswap/MainActivity.java
@@ -1,9 +1,11 @@
 package com.example.aounl.comicfaceswap;
 
 import android.content.res.AssetManager;
+import android.media.Image;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.util.JsonReader;
+import android.util.Log;
 import android.view.View;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -31,22 +33,23 @@ import org.json.JSONArray;
 import org.json.JSONObject;
 
 public class MainActivity extends AppCompatActivity {
+    private ImageAnalyser imageAnalyser = ImageAnalyser.getInstance();
     private final int PICK_IMAGE = 1;
-    private ProgressDialog detectionProgressDialog;
-    private FaceServiceClient faceServiceClient = new FaceServiceRestClient("0ca0c2e7070f4e85864dbc2ba18c8699");
-    private EmotionServiceClient emotionServiceClient = new EmotionServiceRestClient("29361194922442848f54ab8ec5f65d08");
     private static List<Bitmap> comic_faces = new ArrayList<>();
-    private static JSONObject comicLib = new JSONObject();
-    private static AssetManager assetManager;
+    public static AssetManager assetManager;
     List<RecognizeResult> recos;
     Bitmap bmp = null;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
 
         String[] files;
-        assetManager  = getAssets();
-        recos = new ArrayList<RecognizeResult>();
-            try{
+        imageAnalyser = ImageAnalyser.getInstance();
+        imageAnalyser.assetManager  = getAssets();
+        assetManager = imageAnalyser.assetManager;
+        Log.w("Leo", "onCreate");
+
+
+        try{
             files = assetManager.list("");
 
             InputStream input = assetManager.open("marvel_villains.json");
@@ -55,8 +58,9 @@ public class MainActivity extends AppCompatActivity {
                 IOUtils.copy(input, writer, "UTF-8");
                 String theString = writer.toString();
 
-                comicLib = new JSONObject(theString);
-                JSONArray characters = (JSONArray) comicLib.get("results");
+                imageAnalyser.comicLib = new JSONObject(theString);
+                imageAnalyser.comicCharacters = (JSONArray) imageAnalyser.comicLib.get("results");
+                Log.w("Leo", "characters length is " + imageAnalyser.comicCharacters.length());
 //                System.out.println(((JSONArray) comicLib.get("results")).getJSONObject(0).getJSONObject("emotions").getClass());
 
             } catch(Exception e){
@@ -84,7 +88,7 @@ public class MainActivity extends AppCompatActivity {
             }
         });
 
-        detectionProgressDialog = new ProgressDialog(this);
+        imageAnalyser.detectionProgressDialog = new ProgressDialog(this);
     }
 
     public void takePic(View view){
@@ -94,7 +98,7 @@ public class MainActivity extends AppCompatActivity {
 
     public void analyzeCaptured(View view){
         Bundle extras = getIntent().getExtras();
-        byte[] capturedImage = extras.getByteArray("capturedImage");
+        byte[] capturedImage = imageAnalyser.capturedImage;
         if(capturedImage != null){
             Bitmap bitmap = BitmapFactory.decodeByteArray(capturedImage, 0, capturedImage.length);
 
@@ -168,121 +172,12 @@ public class MainActivity extends AppCompatActivity {
     // Detect faces by uploading face images
 // Frame faces after detection
 
-    private void detectAndFrame(final Bitmap imageBitmap)
-    {
-
-        ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
-        imageBitmap.compress(Bitmap.CompressFormat./*PNG*/JPEG, 100, outputStream);
-        ByteArrayInputStream inputStream =
-                new ByteArrayInputStream(outputStream.toByteArray());
-        AsyncTask<InputStream, String, Face[]> detectTask =
-                new AsyncTask<InputStream, String, Face[]>() {
-                    @Override
-                    protected Face[] doInBackground(InputStream... params) {
-                        try {
-                            FaceServiceClient.FaceAttributeType[] attributes = new FaceServiceClient.FaceAttributeType[]{
-                                    FaceServiceClient.FaceAttributeType.Age,
-                                    FaceServiceClient.FaceAttributeType.Gender,
-                                    FaceServiceClient.FaceAttributeType.HeadPose
-                            };
-                            publishProgress("Detecting...");
-                            System.out.println(" nb of params " + params.length);
-//                            InputStream cloner = params[0]
-
-
-                            ByteArrayOutputStream baos = new ByteArrayOutputStream();
-
-// Fake code simulating the copy
-// You can generally do better with nio if you need...
-// And please, unlike me, do something about the Exceptions :D
-                            // ----------------------
-
-                            //BufferedInputStream bufferedInputStream = new BufferedInputStream(params[0]);
-                           // /*Bitmap */bmp = BitmapFactory.decodeStream(bufferedInputStream);
-                           // bmp = Bitmap.createScaledBitmap(bmp, bmp.getWidth(), bmp.getHeight(), false);
-                           // System.out.println("bytecount of this bitmap" + bmp.getByteCount());
-                          //  ByteArrayOutputStream bos = new ByteArrayOutputStream();
-                          //  bmp.compress(Bitmap.CompressFormat.JPEG/*PNG*/, 0 /*ignored for PNG*/, bos);
-                          //  byte[] bitmapdata = bos.toByteArray();
-                           // ByteArrayInputStream bs = new ByteArrayInputStream(bitmapdata);
-
-                            // ------------------------
-
-
-                            byte[] buffer = new byte[1024];
-                            int len;
-                            /* change params 0 so that it reads from a very low def image */
-                            while ((len = params[0].read(buffer)) > -1 ) {
-                                baos.write(buffer, 0, len);
-                            }
-                            baos.flush();
-
-
-
-// Open new InputStreams using the recorded bytes
-// Can be repeated as many times as you wish
-                            InputStream is1 = new ByteArrayInputStream(baos.toByteArray());
-                            InputStream is2 = new ByteArrayInputStream(baos.toByteArray());
-
-
-                            Face[] result = faceServiceClient.detect(
-                                    is1,
-                                    true,         // returnFaceId
-                                    true,        // returnFaceLandmarks
-                                    attributes           // returnFaceAttributes: a string like "age, gender"
-                            );
-                            is1.close();
-
-                                InputStream input = is2;
-                                recos = emotionServiceClient.recognizeImage( input);
-                            is2.close();
-                            for(RecognizeResult reco : recos){
-                                System.out.println(printScores(reco.scores));
-                            }
-
-
-                            if (result == null)
-                            {
-                                publishProgress("Detection Finished. Nothing detected");
-                                return null;
-                            }
-                            publishProgress(
-                                    String.format("Detection Finished. %d face(s) detected",
-                                            result.length));
-                            return result;
-                        } catch (Exception e) {
-                            System.out.println(" I went into the catch " + e.getMessage());
-
-                            publishProgress("Detection failed");
-                            return null;
-                        }
-                    }
-                    @Override
-                    protected void onPreExecute() {
-                        //TODO: show progress dialog
-                        detectionProgressDialog.show();
-                    }
-                    @Override
-                    protected void onProgressUpdate(String... progress) {
-                        //TODO: update progress
-                        detectionProgressDialog.setMessage(progress[0]);
-                    }
-                    @Override
-                    protected void onPostExecute(Face[] result) {
-                        //TODO: update face frames
-                        detectionProgressDialog.dismiss();
-                        if (result == null) return;
-                        ImageView imageView = (ImageView)findViewById(R.id.imageView1);
-//                        imageView.setImageBitmap(drawFaceRectanglesOnBitmap(imageBitmap, result));
-                        imageView.setImageBitmap(drawFacesOnFaces(/*bmp*/imageBitmap, result, recos));
-                        imageBitmap/*bmp*/.recycle();
-                    }
-                };
-        detectTask.execute(inputStream);
+    private void detectAndFrame(final Bitmap imageBitmap){
+        imageAnalyser.detectAndFrame(imageBitmap);
+        ImageView imageView = (ImageView)findViewById(R.id.imageView1);
+        imageView.setImageBitmap(imageAnalyser.bitmap);
+//        imageView.setImageBitmap(imageBitmap);
     }
-
-
-
 
 
     private Bitmap overlay(Bitmap bmp1, Bitmap bmp2) {
@@ -294,46 +189,7 @@ public class MainActivity extends AppCompatActivity {
     }
 
 
-    private static Bitmap drawFacesOnFaces(Bitmap originalBitmap, Face[] faces, List<RecognizeResult> recos){
-        Bitmap bitmap = originalBitmap.copy(Bitmap.Config.ARGB_8888, true);
-        Canvas canvas = new Canvas(bitmap);
-        int nbOfFaces = recos.size();
-        if(nbOfFaces > 0){
-            for(int i = 0; i< nbOfFaces; i++){
-                FaceRectangle faceRectangle = faces[i].faceRectangle;
-                FaceAttribute attribute = faces[i].faceAttributes;
-                HeadPose headPose = attribute.headPose;
-                float roll = (float)(headPose.roll);
 
-                String imageName = findBestComicFace(faces[i], recos.get(i));
-
-
-                float faceWidth = faceRectangle.width;
-                float faceHeight = faceRectangle.height;
-                Bitmap originalComicFace = null;
-                try{
-                    originalComicFace = BitmapFactory.decodeStream(assetManager.open(imageName));
-                } catch (Exception e){
-                    System.out.println(e.getMessage());
-                }
-                float originalComicFaceWidth = originalComicFace.getWidth();
-                float originalComicFaceHeight = originalComicFace.getHeight();
-                float widthFactor = faceWidth/originalComicFaceWidth;
-                float comicFaceMult = 1.75f;
-                int comicWidth = (int)(originalComicFaceWidth*comicFaceMult*widthFactor);
-                int comicHeight = (int)(originalComicFaceHeight*comicFaceMult*widthFactor);
-                Bitmap comicFace = Bitmap.createScaledBitmap(originalComicFace, comicWidth, comicHeight, false);
-                float centerX = faceRectangle.left + faceRectangle.width/2.f;
-                float centerY = faceRectangle.top + faceRectangle.height/2.f;
-                Matrix matrix = new Matrix();
-                matrix.setTranslate(centerX - comicWidth/2.f, centerY - comicHeight/2.f);
-                matrix.preRotate(roll, comicWidth/2.f, comicHeight/2.f);
-
-                canvas.drawBitmap(comicFace, matrix, null);
-            }
-        }
-        return bitmap;
-    }
 
     private static Bitmap drawFaceRectanglesOnBitmap(Bitmap originalBitmap, Face[] faces) {
         Bitmap bitmap = originalBitmap.copy(Bitmap.Config.ARGB_8888, true);
@@ -380,72 +236,6 @@ public class MainActivity extends AppCompatActivity {
         return super.onOptionsItemSelected(item);
     }
 
-    public static float getCost(Scores emotions, float age, String gender, JSONObject comicChar){
-        float cost = 0.0f;
-
-        try{
-        if(!gender.equals(comicChar.getJSONObject("other").get("gender"))){
-                cost = 100.f;
-            } else{
-
-        cost += Math.pow(Math.abs(emotions.anger - (double)comicChar.getJSONObject("emotions").get("anger") *1.0), 2);
-        cost += Math.pow(Math.abs(emotions.contempt - (double)comicChar.getJSONObject("emotions").get("contempt")*1.0), 2);
-        cost += Math.pow(Math.abs(emotions.disgust - (double)comicChar.getJSONObject("emotions").get("disgust")*1.0), 2);
-        cost += Math.pow(Math.abs(emotions.fear - (double)comicChar.getJSONObject("emotions").get("fear")*1.0), 2);
-        cost += Math.pow(Math.abs(emotions.happiness - (double)comicChar.getJSONObject("emotions").get("happiness")*1.0), 2);
-        cost += Math.pow(Math.abs(emotions.neutral - (double)comicChar.getJSONObject("emotions").get("neutral")*1.0), 2);
-        cost += Math.pow(Math.abs(emotions.sadness - (double)comicChar.getJSONObject("emotions").get("sadness")*1.0), 2);
-        cost += Math.pow(Math.abs(emotions.surprise - (double)comicChar.getJSONObject("emotions").get("surprise")*1.0), 2);
-
-        cost += Math.pow(Math.abs(age - (double)comicChar.getJSONObject("other").get("age"))/50.f, 2);
-        }
-        }catch(Exception e){
-            System.out.println("qwertyuio" + e.getMessage());
-        }
-        try{
-            System.out.println(comicChar.getString("name") + "  costs   " + cost);
-        } catch (Exception e){}
-        return cost;
-    }
-    public static String findBestComicFace(Face face, RecognizeResult reco){
-        Scores scores = reco.scores;
-        float age = (float)face.faceAttributes.age;
-        String gender = face.faceAttributes.gender;
-        String answer = "";
-
-        try{
-            int nbCharacters = comicLib.getJSONArray("results").length();
-            float minCost = Float.MAX_VALUE;
-            for(int i = 0; i < nbCharacters; i++){
-                JSONObject character = comicLib.getJSONArray("results").getJSONObject(i);
-                float thisCost = getCost(scores, age, gender, character);
-                if(thisCost < minCost){
-                    minCost = thisCost;
-                    answer = character.getString("name");
-                }
-            }
-
-        } catch (Exception e){
-            System.out.println(e.getMessage());
-        }
-        return answer;
-    }
-
-    public static String printScores(Scores scores){
-        StringBuilder sb = new StringBuilder();
-        sb.append("---------------------------- \n");
-        sb.append("anger: " + scores.anger + " \n");
-        sb.append("contempt: " + scores.contempt + " \n");
-        sb.append("disgust: " + scores.disgust + " \n");
-        sb.append("fear: " + scores.fear + " \n");
-        sb.append("happiness: " + scores.happiness + " \n");
-        sb.append("neutral: " + scores.neutral + " \n");
-        sb.append("sadness: " + scores.sadness + " \n");
-        sb.append("surprise: " + scores.surprise + " \n");
-        sb.append("---------------------------- \n");
-
-        return sb.toString();
-    }
 
     public static Bitmap rotateImage(Bitmap source, float angle) {
         Matrix matrix = new Matrix();
